ROSS cleanup (#135)
This commit fixes warnings when building with -Wall and -Wextra.  In addition, removes old unused code related to prior pthread versions of ROSS as well as removing the ROSS_MEMORY related code, due to no longer being fully functional.

* fixing warnings when building with Wall, etc

* fixing warnings in buddy.c

* fixing warnings in the rest of ROSS core

* fixing warnings in instrumentation

* fixing warnings with phold

* fixing undefined reference causing linking error

* adding warning flags to Travis builds

* fixing some additional warnings that appear in release builds

* minor changes to travis builds

* removing old, unnecessary pthread related code

* fixing ifdefs for ROSS_MEMORY

* removing tw-memory related stuff for now

* removing some unused stuff and unnecessary function arguments

* lots of documentation for network-mpi and removing some unnecessary stuff

* keeping the ROSS models submodule from being used in Travis tests

* minor fix for C++ Travis build

* adding installing phold to bin

* adding some RPATH commands to CMakeLists.txt for Spack

* removed ROSS-Models and template model submodule

* update readme for submodule details

* rpath in pkg-config

* adding develop branch to Travis safelist

+++ core/tw-util.c
@@ -47,7 +47,7 @@ tw_output(tw_lp *lp, const char *fmt, ...)
     va_start(ap, fmt);
     ret = vsnprintf(out->message, sizeof(out->message), fmt, ap);
     va_end(ap);
-    if (ret >= 0 && ret < sizeof(out->message)) {
+    if (ret >= 0 && (unsigned)ret < sizeof(out->message)) {
         // Should be successful
     }
     else {
@@ -163,7 +163,7 @@ pool_alloc(size_t len)
 	void *r;
 
 	for (p = main_pool; p; p = p->next_pool)
-		if ((p->end_free - p->next_free >= len))
+		if ((unsigned)(p->end_free - p->next_free) >= len)
 			break;
 
 	if (!p) {
